Updated pop_tree() in HForest.cc; Wrote new associated tests; Introduced add_tree() error
Changed HForest's pop_tree() to remove the tree with lowest-value root note, as specified in the assignment directions, and updated the tests in test_hforest.cc associated with the pop_tree(). In doing so, I encountered some very strange behavior from HForest's add_tree() method, which seems to be caused by std::make_heap() sorting the forest's vector incorrectly. I don't know of a solution, however, for now, I'll be working on the HW8 code instead.